+++ src/climatevision/generator/transport2030/energy_demand/air.py
@@ -6,7 +6,7 @@ from ...inputs import Inputs
 from ...utils import div
 from ...transport2018.t18 import T18
 
-from ..transport import Transport
+from .transport import Transport
 
 
 def calc_air_domestic(inputs: Inputs, t18: T18) -> "Transport":
